+++ SeriesGuide/src/main/java/com/battlelancer/seriesguide/util/SeasonTools.java
@@ -17,11 +17,26 @@
 package com.battlelancer.seriesguide.util;
 
 import com.battlelancer.seriesguide.enums.SeasonTags;
+import com.uwetrottmann.seriesguide.R;
+
+import android.content.Context;
+import android.text.TextUtils;
 
 public class SeasonTools {
 
     public static boolean hasSkippedTag(String tags) {
         return SeasonTags.SKIPPED.equals(tags);
+    }
+
+    /**
+     * Builds a localized string like "Season 5" or if the number is 0 "Special Episodes".
+     */
+    public static String getSeasonString(Context context, int seasonNumber) {
+        if (seasonNumber == 0) {
+            return context.getString(R.string.specialseason);
+        } else {
+            return context.getString(R.string.season_number, seasonNumber);
+        }
     }
 
 }